pyproject.toml
[tool.poetry]
name = "image-search"
version = "0.1.0"
description = "A simple Python library for searching images in particular database"
authors = ["Kim"]

[tool.poetry.dependencies]
python = "^3.10"
torch = "^2.2.1"
torchsummary = "^1.5.1"
opencv-python = "^4.9.0.80"
gradio = "^4.25.0"
fire = "^0.5.0"
faiss-cpu = "^1.7.4" 

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"